As noted in 62556220, most of the work for the 64bit port is done, only
polishing/testing is remaining. The 32bit version of the compiler seems
stable, it passes the test suite before and after rebuilding itself
repeatedly.

+++ src/gfxlib2/dos/gfx_dos.c
@@ -670,7 +670,7 @@ void fb_hScreenInfo(int *width, int *height, int *depth, int *refresh)
 	*refresh = fb_dos.refresh;
 }
 
-int fb_hGetWindowHandle(void)
+ssize_t fb_hGetWindowHandle(void)
 {
 	return 0;
 }
