+++ src/rtlib/dev_file_readline_wstr.c
@@ -2,7 +2,7 @@
 
 #include "fb.h"
 
-int fb_DevFileReadLineWstr( FB_FILE *handle, FB_WCHAR *dst, int dst_chars )
+int fb_DevFileReadLineWstr( FB_FILE *handle, FB_WCHAR *dst, ssize_t dst_chars )
 {
     int res;
     FILE *fp;
