+++ src/rtlib/file_get_wstr.c
@@ -2,8 +2,14 @@
 
 #include "fb.h"
 
-/*:::::*/
-int fb_FileGetWstrEx( FB_FILE *handle, fb_off_t pos, FB_WCHAR *dst, int dst_chars, size_t *bytesread )
+int fb_FileGetWstrEx
+	(
+		FB_FILE *handle,
+		fb_off_t pos,
+		FB_WCHAR *dst,
+		ssize_t dst_chars,
+		size_t *bytesread
+	)
 {
     int res;
 
@@ -32,26 +38,22 @@ int fb_FileGetWstrEx( FB_FILE *handle, fb_off_t pos, FB_WCHAR *dst, int dst_char
 	return res;
 }
 
-/*:::::*/
-FBCALL int fb_FileGetWstr( int fnum, long pos, FB_WCHAR *dst, int dst_chars )
+FBCALL int fb_FileGetWstr( int fnum, int pos, FB_WCHAR *dst, ssize_t dst_chars )
 {
 	return fb_FileGetWstrEx( FB_FILE_TO_HANDLE(fnum), pos, dst, dst_chars, NULL );
 }
 
-/*:::::*/
-FBCALL int fb_FileGetWstrLarge( int fnum, long long pos, FB_WCHAR *dst, int dst_chars )
+FBCALL int fb_FileGetWstrLarge( int fnum, long long pos, FB_WCHAR *dst, ssize_t dst_chars )
 {
 	return fb_FileGetWstrEx( FB_FILE_TO_HANDLE(fnum), pos, dst, dst_chars, NULL );
 }
 
-/*:::::*/
-FBCALL int fb_FileGetWstrIOB( int fnum, long pos, FB_WCHAR *dst, int dst_chars, unsigned int *bytesread )
+FBCALL int fb_FileGetWstrIOB( int fnum, int pos, FB_WCHAR *dst, ssize_t dst_chars, size_t *bytesread )
 {
 	return fb_FileGetWstrEx( FB_FILE_TO_HANDLE(fnum), pos, dst, dst_chars, bytesread );
 }
 
-/*:::::*/
-FBCALL int fb_FileGetWstrLargeIOB( int fnum, long long pos, FB_WCHAR *dst, int dst_chars, unsigned int *bytesread )
+FBCALL int fb_FileGetWstrLargeIOB( int fnum, long long pos, FB_WCHAR *dst, ssize_t dst_chars, size_t *bytesread )
 {
 	return fb_FileGetWstrEx( FB_FILE_TO_HANDLE(fnum), pos, dst, dst_chars, bytesread );
 }
