+++ src/rtlib/file_input_str.c
@@ -2,11 +2,10 @@
 
 #include "fb.h"
 
-/*:::::*/
-FBCALL int fb_InputString( void *dst, int strlen, int fillrem )
+FBCALL int fb_InputString( void *dst, ssize_t strlen, int fillrem )
 {
     char buffer[FB_INPUT_MAXSTRINGLEN+1];
-    int isfp;
+	int isfp;
 
 	fb_FileInputNextToken( buffer, FB_INPUT_MAXSTRINGLEN, TRUE, &isfp );
