+++ src/rtlib/linux/sys_fmem.c
@@ -3,7 +3,7 @@
 #include "../fb.h"
 #include <sys/sysinfo.h>
 
-FBCALL unsigned int fb_GetMemAvail( int mode )
+FBCALL size_t fb_GetMemAvail( int mode )
 {
 	return get_avphys_pages() * sysconf(_SC_PAGE_SIZE);
 }
